feat: update molecule verify.yml files
This commit updates the `verify.yml` files in the `molecule` directory. The changes include using built-in Ansible modules for failing and waiting, as well as updating the state of netdata service to either started or stopped depending on the scenario. Additionally, it checks if port 19999 is open or closed based on its state.