feat: add support for new OS versions
This commit adds support for new Ubuntu, Debian, Fedora and Amazon Linux versions. Specifically, "jammy" and "focal" for Ubuntu, "bookworm", "bullseye", and "buster" for Debian, version 38, 37 and 36 for Fedora, version 15.4 and 15.3 for opensuse and version 2023 for Amazon Linux are now supported. Additionally, the netdata tag has been added to the galaxy tags list.