feat: add netdata service verification
This commit adds a check for the netdata service to ensure it is enabled and fails if it is disabled. Additionally, it includes a task to ensure that the postfix service is enabled on boot and started.

The changes are made in `molecule/cloud/verify.yml` and `molecule/local/verify.yml`, as well as `tasks/main.yml`.