feat: add static build removal process
This commit refactors the code to improve the process of removing netdata static builds. It introduces a new task file, `remove-static-build.yml`, which checks for the existence of `/opt/netdata` directory and removes any static builds if found. The existing tasks have been updated to include this new task when necessary.

The changes aim to enhance the clarity and efficiency of removing netdata static builds ensuring only native builds exist