feat: add apt cache update step in static-build prepare.yml
This commit adds a new task to update the apt cache in the `prepare.yml` file for the static-build molecule scenario. The task uses the `ansible.builtin.apt` module with `update_cache` set to true and `cache_valid_time` set to 3600. This task is only executed when the target system's OS family is Debian. Additionally, this commit also includes an installation of Curl using the `ansible.builtin.package` module.